requirement.txt
Naked
beautifulsoup4
request